#79 refactor: 리뷰 도메인 분리
