+++ src/main/kotlin/com/psr/psr/order/dto/OrderAssembler.kt
@@ -2,7 +2,7 @@ package com.psr.psr.order.dto
 
 import com.psr.psr.global.Constant.OrderType.OrderType.SELL
 import com.psr.psr.order.entity.Order
-import com.psr.psr.product.entity.product.Product
+import com.psr.psr.product.entity.Product
 import com.psr.psr.user.entity.User
 import org.springframework.stereotype.Component
 import java.time.format.DateTimeFormatter
