+++ src/main/kotlin/com/psr/psr/order/entity/Order.kt
@@ -2,7 +2,7 @@ package com.psr.psr.order.entity
 
 import com.psr.psr.global.entity.BaseEntity
 import com.psr.psr.order.dto.OrderReq
-import com.psr.psr.product.entity.product.Product
+import com.psr.psr.product.entity.Product
 import com.psr.psr.user.entity.User
 import jakarta.persistence.*
 import org.jetbrains.annotations.NotNull
