+++ src/main/kotlin/com/psr/psr/product/entity/ProductLike.kt
@@ -1,4 +1,4 @@
-package com.psr.psr.product.entity.product
+package com.psr.psr.product.entity
 
 import com.psr.psr.global.entity.BaseEntity
 import com.psr.psr.user.entity.User
@@ -6,14 +6,14 @@ import jakarta.persistence.*
 
 @Entity
 data class ProductLike(
-        @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
         var id: Long,
 
-        @ManyToOne
+    @ManyToOne
         @JoinColumn(nullable = false, name = "product_id")
         var product: Product,
 
-        @ManyToOne
+    @ManyToOne
         @JoinColumn(nullable = false, name = "user_id")
         var user: User
