+++ src/main/kotlin/com/psr/psr/product/entity/ProductReport.kt
@@ -1,25 +1,25 @@
-package com.psr.psr.product.entity.product
+package com.psr.psr.product.entity
 
 import com.psr.psr.global.entity.BaseEntity
-import com.psr.psr.product.entity.ReportCategory
+import com.psr.psr.global.entity.ReportCategory
 import com.psr.psr.user.entity.User
 import jakarta.persistence.*
 import org.jetbrains.annotations.NotNull
 
 @Entity
 data class ProductReport(
-        @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
         var id: Long,
 
-        @ManyToOne
+    @ManyToOne
         @JoinColumn(nullable = false, name = "product_id")
         var product: Product,
 
-        @ManyToOne
+    @ManyToOne
         @JoinColumn(nullable = false, name = "user_id")
         var user: User,
 
-        @NotNull
+    @NotNull
         @Enumerated(EnumType.STRING)
         var category: ReportCategory
