+++ src/main/kotlin/com/psr/psr/product/repository/ProductImgRepository.kt
@@ -1,8 +1,8 @@
-package com.psr.psr.product.repository.product
+package com.psr.psr.product.repository
 
 
-import com.psr.psr.product.entity.product.Product
-import com.psr.psr.product.entity.product.ProductImg
+import com.psr.psr.product.entity.Product
+import com.psr.psr.product.entity.ProductImg
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
