+++ src/main/kotlin/com/psr/psr/product/repository/ProductLikeRepository.kt
@@ -1,6 +1,6 @@
-package com.psr.psr.product.repository.product
+package com.psr.psr.product.repository
 
-import com.psr.psr.product.entity.product.ProductLike
+import com.psr.psr.product.entity.ProductLike
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
