+++ src/main/kotlin/com/psr/psr/product/repository/ProductReportRepository.kt
@@ -1,6 +1,6 @@
-package com.psr.psr.product.repository.product
+package com.psr.psr.product.repository
 
-import com.psr.psr.product.entity.product.ProductReport
+import com.psr.psr.product.entity.ProductReport
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
