+++ src/main/kotlin/com/psr/psr/review/entity/Review.kt
@@ -1,7 +1,7 @@
-package com.psr.psr.product.entity.review
+package com.psr.psr.review.entity
 
 import com.psr.psr.global.entity.BaseEntity
-import com.psr.psr.product.entity.product.Product
+import com.psr.psr.product.entity.Product
 import com.psr.psr.user.entity.User
 import jakarta.persistence.*
 import org.jetbrains.annotations.NotNull
