+++ src/main/kotlin/com/psr/psr/review/entity/ReviewReport.kt
@@ -1,25 +1,25 @@
-package com.psr.psr.product.entity.review
+package com.psr.psr.review.entity
 
 import com.psr.psr.global.entity.BaseEntity
-import com.psr.psr.product.entity.ReportCategory
+import com.psr.psr.global.entity.ReportCategory
 import com.psr.psr.user.entity.User
 import jakarta.persistence.*
 import org.jetbrains.annotations.NotNull
 
 @Entity
 data class ReviewReport(
-        @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
         var id: Long,
 
-        @ManyToOne
+    @ManyToOne
         @JoinColumn(nullable = false, name = "review_id")
         var review: Review,
 
-        @ManyToOne
+    @ManyToOne
         @JoinColumn(nullable = false, name = "user_id")
         var user: User,
 
-        @NotNull
+    @NotNull
         @Enumerated(EnumType.STRING)
         var category: ReportCategory
