+++ src/main/kotlin/com/psr/psr/review/repository/ReviewRepository.kt
@@ -1,6 +1,6 @@
-package com.psr.psr.product.repository.review
+package com.psr.psr.review.repository
 
-import com.psr.psr.product.entity.review.Review
+import com.psr.psr.review.entity.Review
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
