+++ src/main/kotlin/com/psr/psr/review/service/ProductService.kt
@@ -1,0 +1,11 @@
+package com.psr.psr.review.service
+
+import com.psr.psr.review.repository.ReviewRepository
+import org.springframework.stereotype.Service
+
+@Service
+class ProductService(
+    private val reviewRepository: ReviewRepository
+) {
+
+}
